Move output/mode util functions to output/mode/modeutil
