+++ libbeat/outputs/mode/mode.go
@@ -82,8 +82,6 @@ type AsyncProtocolClient interface {
 	AsyncPublishEvent(cb func(error), event common.MapStr) error
 }
 
-type ClientFactory func(host string) (ProtocolClient, error)
-
 var (
 	// ErrTempBulkFailure indicates PublishEvents fail temporary to retry.
 	ErrTempBulkFailure = errors.New("temporary bulk send failure")
@@ -93,118 +91,6 @@ var (
 	debug = logp.MakeDebug("output")
 )
 
-func NewConnectionMode(
-	clients []ProtocolClient,
-	failover bool,
-	maxAttempts int,
-	waitRetry, timeout, maxWaitRetry time.Duration,
-) (ConnectionMode, error) {
-	if failover {
-		clients = NewFailoverClient(clients)
-	}
-
-	if len(clients) == 1 {
-		return NewSingleConnectionMode(clients[0], maxAttempts,
-			waitRetry, timeout, maxWaitRetry)
-	}
-	return NewLoadBalancerMode(clients, maxAttempts,
-		waitRetry, timeout, maxWaitRetry)
-}
-
-func NewAsyncConnectionMode(
-	clients []AsyncProtocolClient,
-	failover bool,
-	maxAttempts int,
-	waitRetry, timeout, maxWaitRetry time.Duration,
-) (ConnectionMode, error) {
-	if failover {
-		clients = NewAsyncFailoverClient(clients)
-	}
-	return NewAsyncLoadBalancerMode(clients, maxAttempts,
-		waitRetry, timeout, maxWaitRetry)
-}
-
-// MakeClients will create a list from of ProtocolClient instances from
-// outputer configuration host list and client factory function.
-func MakeClients(
-	config *common.Config,
-	newClient ClientFactory,
-) ([]ProtocolClient, error) {
-	hosts, err := ReadHostList(config)
-	if err != nil {
-		return nil, err
-	}
-	if len(hosts) == 0 {
-		return nil, ErrNoHostsConfigured
-	}
-
-	clients := make([]ProtocolClient, 0, len(hosts))
-	for _, host := range hosts {
-		client, err := newClient(host)
-		if err != nil {
-			// on error destroy all client instance created
-			for _, client := range clients {
-				_ = client.Close() // ignore error
-			}
-			return nil, err
-		}
-		clients = append(clients, client)
-	}
-	return clients, nil
-}
-
-func MakeAsyncClients(
-	config *common.Config,
-	newClient func(string) (AsyncProtocolClient, error),
-) ([]AsyncProtocolClient, error) {
-	hosts, err := ReadHostList(config)
-	if err != nil {
-		return nil, err
-	}
-	if len(hosts) == 0 {
-		return nil, ErrNoHostsConfigured
-	}
-
-	clients := make([]AsyncProtocolClient, 0, len(hosts))
-	for _, host := range hosts {
-		client, err := newClient(host)
-		if err != nil {
-			// on error destroy all client instance created
-			for _, client := range clients {
-				_ = client.Close() // ignore error
-			}
-			return nil, err
-		}
-		clients = append(clients, client)
-	}
-	return clients, nil
-}
-
-func ReadHostList(cfg *common.Config) ([]string, error) {
-	config := struct {
-		Hosts  []string `config:"hosts"`
-		Worker int      `config:"worker"`
-	}{
-		Worker: 1,
-	}
-
-	err := cfg.Unpack(&config)
-	if err != nil {
-		return nil, err
-	}
-
-	lst := config.Hosts
-	if len(lst) == 0 || config.Worker <= 1 {
-		return lst, nil
-	}
-
-	// duplicate entries config.Workers times
-	hosts := make([]string, 0, len(lst)*config.Worker)
-	for _, entry := range lst {
-		for i := 0; i < config.Worker; i++ {
-			hosts = append(hosts, entry)
-		}
-	}
-
-	return hosts, nil
+func Dropped(i int) {
+	messagesDropped.Add(int64(i))
 }
